mbta_info/requirements.txt
alembic==1.4.0
amqp==2.5.2
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
blessings==1.7
bpython==0.18
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==5.1
curtsies==0.3.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
flower==0.9.4
GeoAlchemy2==0.7.0
greenlet==0.4.15
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-enum==1.5.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycountry==19.8.18
Pygments==2.5.2
pyparsing==2.4.7
pytest==5.4.1
pytest-mock==3.1.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
sqlalchemy-stubs==0.3
toml==0.10.0
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==3.1.0
